bbc/requirements.txt
Django==4.1.1
gunicorn==20.0.4
psycopg2-binary==2.9.1
djangorestframework==3.14.0
python-telegram-bot==20.2